park-finder-camera/requirements.txt
Flask
flask-cors
python-dotenv
pymongo
pymongo[srv]
requests
# opencv-python
# opencv-python-headless==4.5.3.56
opencv-contrib-python==4.5.5.62
opencv-python==4.6.0.66
opencv-python-headless<4.3
Pillow
requests
numpy
